+++ Ruddarr/Models/MediaFile.swift
@@ -17,6 +17,23 @@ struct MediaFile: Identifiable, Equatable, Codable {
     let seriesId: Series.ID?
     let episodeReleaseType: EpisodeReleaseType?
 
+    init(from decoder: any Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+
+        id = try container.decode(Int.self, forKey: .id)
+        size = try container.decode(Int.self, forKey: .size)
+        relativePath = try container.decodeIfPresent(String.self, forKey: .relativePath)
+        dateAdded = try container.decode(Date.self, forKey: .dateAdded)
+        mediaInfo = try container.decodeIfPresent(FileMediaInfo.self, forKey: .mediaInfo)
+        quality = try container.decode(MediaQuality.self, forKey: .quality)
+        customFormats = try container.decodeIfPresent([MediaCustomFormat].self, forKey: .customFormats)
+        customFormatScore = try container.decodeIfPresent(Int.self, forKey: .customFormatScore)
+        seriesId = try container.decodeIfPresent(Series.ID.self, forKey: .seriesId)
+        episodeReleaseType = try container.decodeIfPresent(EpisodeReleaseType.self, forKey: .episodeReleaseType)
+
+        languages = try container.decodeLossyArrayIfPresent([MediaLanguage].self, forKey: .languages)
+    }
+
     var filenameLabel: String {
         relativePath?.components(separatedBy: "/").last?.breakable() ?? "--"
     }